chore: update  dockerfile and scripts

+++ crm_reload.sh
@@ -1,3 +1,5 @@
+local expres="${1:-false}"    # Default value is false
+
 echo "Clearing code dump & swagger..."
 php composer.phar dump-autoload
 php artisan l5-swagger:generate
@@ -27,16 +29,18 @@ echo "Generating JWT Secret..."
 php artisan jwt:secret --force --quiet
 echo "Generated JWT Secret Key"
 
-echo "Creating DB..."
-php artisan migrate:refresh
-echo "DB created."
+if "$expres"; then
+    echo "Creating DB..."
+    php artisan migrate:refresh
+    echo "DB created."
 
-echo "Creating meta and dummy data ..."
-php artisan module:seed Core
-php artisan module:seed Contact
-php artisan module:seed User
-php artisan module:seed Preference
-php artisan module:seed Subscription
-php artisan module:seed Account
-php artisan module:seed ServiceRequest
-echo "Data created..."
+    echo "Creating meta and dummy data ..."
+    php artisan module:seed Core
+    php artisan module:seed Contact
+    php artisan module:seed User
+    php artisan module:seed Preference
+    php artisan module:seed Subscription
+    php artisan module:seed Account
+    php artisan module:seed ServiceRequest
+    echo "Data created..."
+fi
